requirements.txt
amqp==2.5.1
aniso8601==8.0.0
attrs==19.3.0
billiard==3.6.1.0
boto3==1.10.14
botocore==1.13.14
celery==4.3.0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
docker==4.1.0
docutils==0.15.2
Flask==1.1.1
Flask-Cors==3.0.8
flask-restplus==0.13.0
gitdb2==2.0.6
gitlab==1.0.2
GitPython==3.0.4
grpcio==1.25.0
grpcio-tools==1.25.0
idna==2.8
importlib-metadata==0.23
itsdangerous==1.1.0
Jinja2==2.10.3
jmespath==0.9.4
jsonschema==3.1.1
kombu==4.6.5
lxml==4.4.1
MarkupSafe==1.1.1
more-itertools==7.2.0
protobuf==3.10.0
pyhelm==2.14.5
pyrsistent==0.15.5
python-dateutil==2.8.0
pytz==2019.3
PyYAML==5.1.2
requests==2.22.0
s3transfer==0.2.1
six==1.13.0
smmap2==2.0.5
supermutes==0.2.5
urllib3==1.25.6
vine==1.3.0
websocket-client==0.56.0
Werkzeug==0.16.0
zipp==0.6.0
